+++ .bud/Build.cs
@@ -7,7 +7,6 @@ using Bud.Resources;
 public class Build : IBuild {
   public Settings Setup(Settings settings, string baseDir) {
     var budCore = new Project("Bud.Core",
-                              Path.Combine(baseDir, "Core"),
                               Res.Main(),
                               Cs.Dll(CSharpKeys.RootNamespace.Modify("Bud"),
                                      Cs.Dependency("Microsoft.Bcl.Immutable"),
@@ -19,23 +18,19 @@ public class Build : IBuild {
                                       Cs.Dependency("NUnit")));
 
     var bud = new Project("bud",
-                          Path.Combine(baseDir, "bud"),
                           Cs.Exe(CSharpKeys.RootNamespace.Modify("Bud"),
                                  Cs.Dependency("Bud.Core"),
                                  Cs.Dependency("CommandLineParser")));
 
     var budTest = new Project("Bud.Test",
-                              Path.Combine(baseDir, "Test"),
                               Cs.Dll(Cs.Dependency("Bud.Core"),
                                      Cs.Dependency("NUnit")));
 
     var budSystemTests = new Project("Bud.SystemTests",
-                                     Path.Combine(baseDir, "SystemTests"),
                                      Cs.Test(Cs.Dependency("Bud.Test"),
                                              NUnitPlugin.NUnitArgs.Modify(list => list.Add("/noshadow"))));
 
     var budExamplesSnippets = new Project("Bud.Examples.Snippets",
-                                          Path.Combine(baseDir, "Bud.Examples.Snippets"),
                                           Cs.Dll(Cs.Dependency("Bud.Core")));
 
     return settings.AddGlobally(CSharpKeys.TargetFramework.Init(Framework.Net46))