+++ Bud.Core/src/main/cs/BudVersion.cs
@@ -1,5 +1,5 @@
 namespace Bud {
   public static class BudVersion {
-    public const string Current = "0.1.2";
+    public const string Current = "0.1.3-dev";
   }
 }